Refactor BridgeIn component and add RecentTransactions display
- Remove unused props and imports from BridgeIn component to simplify code.
- Introduce RecentTransactions component to display user's recent deposit and withdrawal activities.
- Update DepositsPage to integrate RecentTransactions, enhancing user experience by showing transaction history.

Made-with: AI called Ariya
